rename package to 'testing'
For easier differentiation from the gradle ip.test directory
